+++ sdk/python/inkbox/client.py
@@ -28,6 +28,7 @@ from inkbox.a2a.invitations import (
 from inkbox.agent_identity import AgentIdentity
 from inkbox.api_keys.resources.api_keys import ApiKeysResource
 from inkbox.contacts.resources.contacts import ContactsResource
+from inkbox.companion import CompanionResource
 from inkbox.notes.resources.notes import NotesResource
 from inkbox.agent_signup.types import (
     AgentSignupResponse,
@@ -312,6 +313,7 @@ class Inkbox:
         self._a2a_invitations = A2AInvitationsResource(self._api_http, self._base_url)
 
         self._contacts = ContactsResource(self._contacts_http)
+        self.companion = CompanionResource(self._api_http)
         self._notes = NotesResource(self._contacts_http)
 
         self._tunnels = TunnelsResource(self._api_http, inkbox=self)